enum

+++ lib/logitech_receiver/base.py
@@ -343,16 +343,16 @@ class _HIDPP_Notification(NamedTuple):
 
 
 class ReadException(Exception):
-    def __init__(self, protocol: int, code: int):
+    def __init__(self, protocol: float, code: int):
         self.protocol_version = protocol
         self.code = code
         super().__init__()
 
     def error(self):
-        return {1: _hidpp10, 2: _hidpp20}[self.protocol_version].ERROR[self.code]
+        return {1.0: _hidpp10, 2.0: _hidpp20}[self.protocol_version].ERROR[self.code]
 
     def __str__(self):
-        return f'HIDPP{self.protocol_version}0: {self.error()}'
+        return f"HIDPP{self.protocol_version}0: {self.error()}"
 
 
 #
@@ -412,11 +412,11 @@ class Request:
     def handle_reply(self, reply: RawPacket):
         exception = self.v1_exception(reply)
         if exception:
-            raise ReadException(protocol=1, code=exception)
+            raise ReadException(protocol=1.0, code=exception)
 
         exception = self.v2_exception(reply)
         if exception:
-            raise ReadException(protocol=2, code=exception)
+            raise ReadException(protocol=2.0, code=exception)
 
         return self.get_data(reply)
